move diff_components from Measure to util file and call in create_condition_and_components
