+++ measures/models.py
@@ -13,7 +13,6 @@ from common.models.managers import TrackedModelManager
 from common.models.mixins.validity import ValidityMixin
 from common.util import TaricDateRange
 from common.util import classproperty
-from common.validators import UpdateType
 from footnotes import validators as footnote_validators
 from measures import business_rules
 from measures import validators
@@ -21,7 +20,6 @@ from measures.querysets import MeasureConditionQuerySet
 from measures.querysets import MeasuresQuerySet
 from quotas import business_rules as quotas_business_rules
 from quotas.validators import quota_order_number_validator
-from workbaskets.models import WorkBasket
 
 
 class MeasureTypeSeries(TrackedModel, ValidityMixin):
@@ -606,54 +604,6 @@ class Measure(TrackedModel, ValidityMixin):
             .filter(condition__dependent_measure__sid=self.sid)
             .exists()
         )
-
-    def diff_components(
-        self,
-        duty_sentence: str,
-        start_date: date,
-        workbasket: WorkBasket,
-    ):
-        from measures.parsers import DutySentenceParser
-
-        parser = DutySentenceParser.get(
-            start_date,
-        )
-
-        new_components = parser.parse(duty_sentence)
-        old_components = self.components.approved_up_to_transaction(
-            workbasket.current_transaction,
-        )
-        new_by_id = {c.duty_expression.id: c for c in new_components}
-        old_by_id = {c.duty_expression.id: c for c in old_components}
-        all_ids = set(new_by_id.keys()) | set(old_by_id.keys())
-        update_transaction = None
-        for id in all_ids:
-            new = new_by_id.get(id)
-            old = old_by_id.get(id)
-            if new and old:
-                # Component is having amount/unit changed – UPDATE it
-                new.update_type = UpdateType.UPDATE
-                new.version_group = old.version_group
-                new.component_measure = self
-                if not update_transaction:
-                    update_transaction = workbasket.new_transaction()
-                new.transaction = update_transaction
-                new.save()
-
-            elif new:
-                # Component exists only in new set - CREATE it
-                new.update_type = UpdateType.CREATE
-                new.component_measure = self
-                new.transaction = workbasket.new_transaction()
-                new.save()
-
-            elif old:
-                # Component exists only in old set – DELETE it
-                old = old.new_version(
-                    workbasket,
-                    update_type=UpdateType.DELETE,
-                    transaction=workbasket.new_transaction(),
-                )
 
 
 class MeasureComponent(TrackedModel):
