+++ measures/tests/test_models.py
@@ -5,7 +5,6 @@ import pytest
 from common.tests import factories
 from common.validators import UpdateType
 from measures.models import Measure
-from measures.models import MeasureComponent
 
 pytestmark = pytest.mark.django_db
 
@@ -301,127 +300,3 @@ def test_effective_valid_between(measure_regulation, measure_dates, date_ranges)
     )
     assert measure.effective_valid_between == expected_dates
     assert measure.effective_end_date == expected_dates.upper
-
-
-def test_diff_components_update(
-    workbasket,
-    duty_sentence_parser,
-    percent_or_amount,
-):
-    original_component = factories.MeasureComponentFactory.create(
-        duty_amount=9.000,
-        duty_expression=percent_or_amount,
-    )
-    original_component.component_measure.diff_components(
-        "8.000%",
-        original_component.component_measure.valid_between.lower,
-        workbasket,
-    )
-    components = (
-        original_component.component_measure.components.approved_up_to_transaction(
-            workbasket.current_transaction,
-        )
-    )
-
-    assert components.count() == 1
-
-    new_component = components.last()
-
-    assert new_component.update_type == UpdateType.UPDATE
-    assert new_component.version_group == original_component.version_group
-    assert new_component.component_measure == original_component.component_measure
-    assert new_component.transaction == workbasket.current_transaction
-    assert new_component.duty_amount == 8.000
-
-
-def test_diff_components_update_multiple(
-    workbasket,
-    duty_sentence_parser,
-    percent_or_amount,
-    plus_percent_or_amount,
-    monetary_units,
-    measurement_units,
-):
-    component_1 = factories.MeasureComponentFactory.create(
-        duty_amount=12.000,
-        duty_expression=percent_or_amount,
-    )
-    component_2 = factories.MeasureComponentFactory.create(
-        component_measure=component_1.component_measure,
-        duty_amount=253.000,
-        duty_expression=plus_percent_or_amount,
-        monetary_unit=monetary_units["GBP"],
-        component_measurement__measurement_unit=measurement_units[1],
-    )
-    component_2.component_measure.diff_components(
-        "13.000% + 254.000 GBP / 100 kg",
-        component_1.component_measure.valid_between.lower,
-        workbasket,
-    )
-    components = component_1.component_measure.components.approved_up_to_transaction(
-        workbasket.current_transaction,
-    )
-
-    assert components.count() == 2
-
-    first = components.filter(
-        duty_expression__sid=component_1.duty_expression.sid,
-    ).first()
-    second = components.filter(
-        duty_expression__sid=component_2.duty_expression.sid,
-    ).first()
-
-    assert components.count() == 2
-    assert first.duty_amount == 13.000
-    assert second.duty_amount == 254.000
-    assert components.first().transaction == components.last().transaction
-
-
-def test_diff_components_create(workbasket, duty_sentence_parser):
-    measure = factories.MeasureFactory.create()
-    measure.diff_components(
-        "8.000%",
-        measure.valid_between.lower,
-        workbasket,
-    )
-    components = measure.components.approved_up_to_transaction(
-        workbasket.current_transaction,
-    )
-
-    assert components.count() == 1
-
-    new_component = components.last()
-
-    assert new_component.update_type == UpdateType.CREATE
-    assert new_component.component_measure == measure
-    assert new_component.transaction == workbasket.current_transaction
-    assert new_component.duty_amount == 8.000
-
-
-def test_diff_components_delete(
-    workbasket,
-    duty_sentence_parser,
-    percent_or_amount,
-):
-    component = factories.MeasureComponentFactory.create(
-        duty_amount=9.000,
-        duty_expression=percent_or_amount,
-    )
-    component.component_measure.diff_components(
-        "",
-        component.component_measure.valid_between.lower,
-        workbasket,
-    )
-    components = component.component_measure.components.approved_up_to_transaction(
-        workbasket.current_transaction,
-    )
-
-    assert components.count() == 0
-
-    deleted = MeasureComponent.objects.filter(
-        component_measure=component.component_measure,
-        update_type=UpdateType.DELETE,
-    )
-
-    assert deleted.exists()
-    assert deleted.first().transaction == workbasket.current_transaction
